Use UInt64 for balance constants to avoid repeatedly creating new UInt64 instances for them. (#3367)

+++ ethereum/datastructures/src/main/java/tech/pegasys/teku/datastructures/util/ValidatorsUtil.java
@@ -50,7 +50,7 @@ public class ValidatorsUtil {
    */
   public static boolean is_eligible_for_activation_queue(Validator validator) {
     return validator.getActivation_eligibility_epoch().equals(Constants.FAR_FUTURE_EPOCH)
-        && validator.getEffective_balance().equals(UInt64.valueOf(Constants.MAX_EFFECTIVE_BALANCE));
+        && validator.getEffective_balance().equals(Constants.MAX_EFFECTIVE_BALANCE);
   }
 
   /**
